requirements.txt
streamlit>=1.40.0
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
duckduckgo-search>=5.0.0
pytest==8.0.0